Finished unit tests for DAO layer

+++ dao/src/main/java/com/service/dao/currency/CurrencyDAO.java
@@ -2,9 +2,10 @@ package com.service.dao.currency;
 
 import com.service.common.enums.Currency;
 import com.service.common.exceptions.InvalidCurrencyConversionException;
+import java.math.BigDecimal;
 
 public interface CurrencyDAO {
 
-    Double convertCurrency(Currency senderCurrency, Currency receiverCurrency, Double amount) throws InvalidCurrencyConversionException;
+    BigDecimal convertCurrency(Currency senderCurrency, Currency receiverCurrency, BigDecimal amount) throws InvalidCurrencyConversionException;
 
 }
